+++ pythermalcomfort/models/at.py
@@ -3,8 +3,7 @@ from typing import Union, List
 
 import numpy as np
 
-from pythermalcomfort.psychrometrics import psy_ta_rh
-from pythermalcomfort.utilities import BaseInputs
+from pythermalcomfort.utilities import BaseInputs, psy_ta_rh
 
 
 @dataclass(frozen=True)